--- api/Dockerfile
@@ -1,17 +1,0 @@
-FROM oven/bun
-
-WORKDIR /app
-
-COPY package.json .
-COPY bun.lockb .
-
-RUN bun install --production
-
-COPY src src
-COPY tsconfig.json .
-# COPY public public
-
-ENV NODE_ENV production
-CMD ["bun", "src/index.ts"]
-
-EXPOSE 3000
