--- api/context.ts
@@ -1,17 +1,0 @@
-// From https://www.prisma.io/docs/orm/prisma-client/testing/unit-testing#dependency-injection
-import { PrismaClient } from '@prisma/client'
-import { mockDeep, DeepMockProxy } from 'jest-mock-extended'
-
-export type Context = {
-  prisma: PrismaClient
-}
-
-export type MockContext = {
-  prisma: DeepMockProxy<PrismaClient>
-}
-
-export const createMockContext = (): MockContext => {
-  return {
-    prisma: mockDeep<PrismaClient>(),
-  }
-}
