--- api/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "elysia-prisma",
-  "version": "1.0.50",
-  "scripts": {
-    "test": "jest",
-    "dev": "bun run --watch src/index.ts"
-  },
-  "dependencies": {
-    "@elysiajs/cors": "^0.8.0",
-    "@elysiajs/swagger": "^0.8.5",
-    "@prisma/client": "5.9.1",
-    "api": "^6.1.1",
-    "elysia": "latest"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.12",
-    "bun-types": "latest",
-    "jest": "^29.7.0",
-    "jest-mock-extended": "^3.0.5",
-    "prisma": "^5.9.1",
-    "ts-jest": "^29.1.2",
-    "typescript": "^5.3.3"
-  },
-  "module": "src/index.js"
-}
