--- api/prisma/schema.prisma
@@ -1,28 +1,0 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-generator client {
-  provider = "prisma-client-js"
-}
-
-datasource db {
-  provider  = "postgresql"
-  url       = env("DATABASE_URL")
-  directUrl = env("DIRECT_URL")
-}
-
-model Airport {
-  id         Int          @id @default(autoincrement())
-  code       String       @unique
-  departures Connection[] @relation("Departing")
-  arrivals   Connection[] @relation("Arriving")
-}
-
-model Connection {
-  id          Int     @id @default(autoincrement())
-  depart      Airport @relation("Departing", fields: [departureId], references: [id])
-  departureId Int
-  arrive      Airport @relation("Arriving", fields: [arrivalId], references: [id])
-  arrivalId   Int
-  weight      Int
-}
