--- api/tests/processors/requestProcessor.test.ts
@@ -1,15 +1,0 @@
-import { MockContext, Context, createMockContext } from '../../context'
-
-let mockCtx: MockContext
-let ctx: Context
-
-beforeEach(() => {
-  mockCtx = createMockContext()
-  ctx = mockCtx as unknown as Context
-})
-
-describe("Processing of data passed to /request", () => {
-  test.todo("Test endpoint parameters")
-  test.todo("Test error handling")
-  test.todo("Test prisma integration")
-})
