+++ app/components/FlightRequestForm.tsx
@@ -11,38 +11,16 @@ import {
 } from "@/components/ui/form"
 import { Popover, PopoverContent, PopoverTrigger } from "./ui/popover"
 
-// import type { API } from "../../api/src/index";
 import { Button } from "@/components/ui/button"
 import { Calendar } from "./ui/calendar"
 import { CalendarIcon } from "lucide-react"
-import { Input } from "@/components/ui/input"
-import { ResponseData } from "../../api/src/clients/Seats.Aero/types"
-import airports from "../lib/airports"
+import { ResponseData } from "@/lib/types"
+import airports from "@/lib/airports"
 import { cn } from "@/lib/utils"
-// import { edenTreaty } from "@elysiajs/eden"
 import { format } from "date-fns"
-import { promises as fs } from 'fs';
-import { toast } from "@/components/ui/use-toast"
 import { useForm } from "react-hook-form"
 import { z } from "zod"
 import { zodResolver } from "@hookform/resolvers/zod"
-
-type Airport = {
-    "icao": string,
-    "iata": string,
-    "name": string,
-    "city": string,
-    "state": string,
-    "country": string,
-    "elevation": number,
-    "lat": number,
-    "lon":  number,
-    "tz": string
-}
-
-type AirportMap = {
-    [key: string]: Airport
-}
 
 const FormSchema = z.object({
     outboundAirportCode: z.string().min(3, {
@@ -58,11 +36,6 @@ const FormSchema = z.object({
     startOutDate: z.date(),
     endOutDate: z.date(),
 })
-
-// const ENDPOINT = process.env.NODE_ENV === "development" ? "http://localhost:4000" : process.env.NEXT_API_URL
-const ENDPOINT = "https://437-webapp.azurewebsites.net"
-
-// const api = edenTreaty<API>(ENDPOINT!)
 
 type Props = {
     setData: (data: ResponseData) => void,
@@ -82,34 +55,29 @@ export function FlightRequestForm(props: Props) {
 
   const onSubmit = (data: z.infer<typeof FormSchema>) => {
     props.setLoading(true)
-    // api.api.request.post({
-    //   origin_airport: data.outboundAirportCode, // TODO: make these names align
-    //   destination_airport: data.inboundAirportCode,
-    //   start_date: data.startOutDate.toISOString().split('T')[0],
-    //   end_date: data.endOutDate.toISOString().split('T')[0]
-    
-    fetch(`${ENDPOINT}/api/request`, {
+
+    fetch('/api/get-flights',{
       method: 'POST',
       headers: {
         'Content-Type': 'application/json'
-        },
-        "mode": "cors",
-        body: JSON.stringify({
-          origin_airport: data.outboundAirportCode, // TODO: make these names align
-          destination_airport: data.inboundAirportCode,
-          start_date: data.startOutDate.toISOString().split('T')[0],
-          end_date: data.endOutDate.toISOString().split('T')[0]
-        })
+      },
+      body: JSON.stringify({
+        outboundAirportCode: data.outboundAirportCode,
+        inboundAirportCode: data.inboundAirportCode,
+        outboundDate: data.startOutDate,
+        inboundDate: data.endOutDate
+      })
     }).then((res) => {
-      if (res.status === 500) {
-        throw new Error("Internal server error")
-      }
-      return res
-    }).then((res) => res.json()).then((res) => {
-      console.log(res)
-      props.setData(res as ResponseData)
-      props.setLoading(false)
-    })
+        if (res.status === 500) {
+            throw new Error("Internal server error")
+        }
+        console.log(res)
+        return res
+    }
+    ).then((res) => res.json()).then((res) => {
+        props.setData(res as ResponseData)
+        props.setLoading(false)
+    }) // TODO: add error catching
   }
   
   return (
